chore: rename TransferCol to TransferColumn
