+++ src/components/TransactionHistoryRows/TransactionDetails/Transfers.tsx
@@ -2,7 +2,7 @@ import { Stack, Tag } from '@chakra-ui/react'
 import type { TxTransfer } from '@shapeshiftoss/chain-adapters'
 import { Text } from 'components/Text'
 
-import { TransferColumn } from './TransferCol'
+import { TransferColumn } from './TransferColumn'
 
 type TransfersProps = {
   compactMode?: boolean
